+++ mobile/android/android-components/buildSrc/src/main/java/Gecko.kt
@@ -6,7 +6,7 @@ internal object GeckoVersions {
     /**
      * GeckoView Nightly Version.
      */
-    const val nightly_version = "70.0.20190809095611"
+    const val nightly_version = "70.0.20190813095543"
 
     /**
      * GeckoView Beta Version.